requirements.txt
rich
InquirerPy
requests
open-interpreter
python-dotenv
